+++ src/main/java/com/example/demo/DTO/MovimientoInventarioResponseDTO.java
@@ -16,7 +16,7 @@ import lombok.NoArgsConstructor;
 
 public class MovimientoInventarioResponseDTO {
     private String productoNombre; 
-    private int precio;
+    private Double precio;
 
     private String categoria; 
     private int cantidad;
